Warnings :   - [ALL] Rajout constructeur + remarques warnings appliquées

+++ robin.cpp
@@ -24,7 +24,7 @@ void exit_handler(int s) {
 
 
 int main(int argc, char** argv) {
-    uint8_t nodeID = read_int_from_file(CFG_NODEID_FNAME);
+    uint8_t nodeID = static_cast<uint8_t>(read_int_from_file(CFG_NODEID_FNAME));
     LOG_F(INFO, "CFG: NodeID=%d", nodeID);
     Tracker tracker(TRACKING_PEER_LOSS_TIMEOUT, TRACKING_PERIOD_CHECK_NODEMAP);
     TrackingServer trackingserver(TRACKING_SERVER_PORT, nodeID, tracker, TRACKING_HEARTBEAT_PERIOD);
